+++ test-relation-properties/Main.hs
@@ -1,0 +1,7 @@
+module Main where
+
+import Test.Hspec
+import RelationPropertiesSpec 
+
+main :: IO ()
+main = hspec spec